requirements.txt
google-generativeai
PyYAML
python-dotenv
httpx
tqdm
google-api-core
google-auth